refactor: split content search view panels
